requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.0
ecdsa==0.18.0
fastapi==0.89.1
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing-extensions==4.4.0
zipp==3.11.0
